favorite

+++ app/Http/Controllers/FavoriteController.php
@@ -12,23 +12,31 @@ class FavoriteController extends Controller
     {
         $request->validate([
             'product_id' => 'required|integer|exists:products,id',
+            'is_favorite' => 'boolean|required'
         ]);
-        Favorite::firstOrCreate(['user_id'=>auth()->user()->id,'product_id'=>$request->product_id],$request->all());
-
-        return response()->json([
-            'status' => 1,
-            'message' => 'Favorite added successfully'
-        ]);
+        if($request->is_favorite==true){
+            Favorite::where('user_id','=',auth()->user()->id)->where('product_id','=',$request->product_id)->delete();
+            return response()->json([
+                'status' => 1,
+                'message' => 'Favorite removed successfully'
+            ]);}
+        {
+            Favorite::firstOrCreate(['user_id' => auth()->user()->id, 'product_id' => $request->product_id], $request->all());
+            return response()->json([
+                'status' => 1,
+                'message' => 'Favorite added successfully'
+            ]);
+        }
     }
-    public function removeFavorite(Request $request)
-    {
-        $request->validate([
-            'product_id' => 'required|integer|exists:products,id',
-        ]);
-        Favorite::where('user_id','=',auth()->user()->id)->where('product_id','=',$request->product_id)->delete();
-        return response()->json([
-            'status' => 1,
-            'message' => 'Favorite removed successfully'
-        ]);
-    }
+//    public function removeFavorite(Request $request)
+//    {
+//        $request->validate([
+//            'product_id' => 'required|integer|exists:products,id',
+//        ]);
+//        Favorite::where('user_id','=',auth()->user()->id)->where('product_id','=',$request->product_id)->delete();
+//        return response()->json([
+//            'status' => 1,
+//            'message' => 'Favorite removed successfully'
+//        ]);
+//    }
 }
